access matchers: extract payload classes into a separate file group
The previous commit removed the tight dependency between the payloads and the
matching engine making it possible to separate them into two file groups and
thus reduce the size the each file.  A caller may eventually want to define
their own payloads instead of using those provided, at which point they don't
need to include the predefined payloads.  Uninline large functions from the
payload that were previously inlined because declared in the header file
together with the template-based matching engine.
