+++ CMakeLists.txt
@@ -76,7 +76,8 @@ add_library(islutils
 
             islutils/matchers.cc
             islutils/parser.cc
-            islutils/builders.cc)
+            islutils/builders.cc
+            islutils/access_patterns.cc)
 target_include_directories(islutils PUBLIC ${CMAKE_SOURCE_DIR})
 target_include_directories(islutils PUBLIC ${isl_INCLUDE} ${pet_INCLUDE})
 target_link_libraries(islutils ${isl_LIB} ${pet_LIB})
